@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.0.RELEASE</version>
+        <version>2.3.3.RELEASE</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>br.com.example.auth.security</groupId>
@@ -15,7 +15,7 @@
     <description>Demo project for Spring Boot with a security layer</description>
 
     <properties>
-        <java.version>11</java.version>
+        <java.version>14</java.version>
         <spring-cloud.version>Hoxton.SR5</spring-cloud.version>
     </properties>
 
@@ -51,19 +51,15 @@
 			<artifactId>h2</artifactId>
 			<scope>runtime</scope>
 		</dependency>
-	    <dependency>
-		    <groupId>io.jsonwebtoken</groupId>
-		    <artifactId>jjwt</artifactId>
-		    <version>0.9.1</version>
-	    </dependency>
+<!--	    <dependency>-->
+<!--		    <groupId>io.jsonwebtoken</groupId>-->
+<!--		    <artifactId>jjwt</artifactId>-->
+<!--		    <version>0.9.1</version>-->
+<!--	    </dependency>-->
 	    <dependency>
 		    <groupId>org.modelmapper</groupId>
 		    <artifactId>modelmapper</artifactId>
 		    <version>2.3.8</version>
-	    </dependency>
-	    <dependency>
-		    <groupId>org.springframework.boot</groupId>
-		    <artifactId>spring-boot-starter-mail</artifactId>
 	    </dependency>
 	    <dependency>
 		    <groupId>org.springframework.boot</groupId>
@@ -80,11 +76,12 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-devtools</artifactId>
-			<scope>runtime</scope>
-		</dependency>
+
+<!--		<dependency>-->
+<!--			<groupId>io.springfox</groupId>-->
+<!--			<artifactId>springfox-swagger2</artifactId>-->
+<!--			<version>3.0.0</version>-->
+<!--		</dependency>-->
 	</dependencies>
 	<dependencyManagement>
 		<dependencies>
